feat(server): migrate /api/agents/avatar to Bun-shape handler (#147 TB2)
